FastAPI/requirements.txt
transformers
numpy
torch
uvicorn
fastapi
tensorflow
pillow
prometheus-fastapi-instrumentator
prometheus_client
psutil